+++ generate_latex.php
@@ -246,7 +246,9 @@ class generate_latex implements WikiIocModel{
             $result = array();
             $this->createLatex($output_filename, DOKU_IOCEXPORTL_LATEX_TMP.$this->tmp_dir, $latex, $result);
         }
-        $this->removeDir(DOKU_IOCEXPORTL_LATEX_TMP.$this->tmp_dir);
+        if(!$conf['plugin']['iocexportl']['saveWorkDir']){
+            $this->removeDir(DOKU_IOCEXPORTL_LATEX_TMP.$this->tmp_dir);
+        }
         if($this->log){
             return $result;
         }
